+++ packages/apollo-server-cloud-function/src/index.ts
@@ -1,0 +1,24 @@
+export {
+  GraphQLUpload,
+  GraphQLOptions,
+  GraphQLExtension,
+  Config,
+  gql,
+  // Errors
+  ApolloError,
+  toApolloError,
+  SyntaxError,
+  ValidationError,
+  AuthenticationError,
+  ForbiddenError,
+  UserInputError,
+  // playground
+  defaultPlaygroundOptions,
+  PlaygroundConfig,
+  PlaygroundRenderPageOptions,
+} from 'apollo-server-core';
+
+export * from 'graphql-tools';
+
+// ApolloServer integration.
+export { ApolloServer, CreateHandlerOptions } from './ApolloServer';